+++ stable_baselines/sac/sac.py
@@ -55,6 +55,9 @@ class SAC(OffPolicyRLModel):
     :param target_entropy: (str or float) target entropy when learning ent_coef (ent_coef = 'auto')
     :param action_noise: (ActionNoise) the action noise type (None by default), this can help
         for hard exploration problem. Cf DDPG for the different action noise type.
+    :param random_exploration: (float) Probability of taken a random action (as in an epsilon-greedy strategy)
+        This is not needed for SAC normally but can help exploring when using HER + SAC.
+        This hack was present in the original OpenAI Baselines repo (DDPG + HER)
     :param verbose: (int) the verbosity level: 0 none, 1 training information, 2 tensorflow debug
     :param tensorboard_log: (str) the log location for tensorboard (if None, no logging)
     :param _init_setup_model: (bool) Whether or not to build the network at the creation of the instance
@@ -67,7 +70,7 @@ class SAC(OffPolicyRLModel):
                  learning_starts=100, train_freq=1, batch_size=64,
                  tau=0.005, ent_coef='auto', target_update_interval=1,
                  gradient_steps=1, target_entropy='auto', action_noise=None,
-                 verbose=0, tensorboard_log=None,
+                 random_exploration=0.0, verbose=0, tensorboard_log=None,
                  _init_setup_model=True, policy_kwargs=None, full_tensorboard_log=False):
 
         super(SAC, self).__init__(policy=policy, env=env, replay_buffer=None, verbose=verbose,
@@ -90,6 +93,7 @@ class SAC(OffPolicyRLModel):
         self.gradient_steps = gradient_steps
         self.gamma = gamma
         self.action_noise = action_noise
+        self.random_exploration = random_exploration
 
         self.value_fn = None
         self.graph = None
@@ -393,8 +397,10 @@ class SAC(OffPolicyRLModel):
 
                 # Before training starts, randomly sample actions
                 # from a uniform distribution for better exploration.
-                # Afterwards, use the learned policy.
-                if self.num_timesteps < self.learning_starts:
+                # Afterwards, use the learned policy
+                # if random_exploration is set to 0 (normal setting)
+                if (self.num_timesteps < self.learning_starts
+                    or np.random.rand() < self.random_exploration):
                     action = self.env.action_space.sample()
                     # No need to rescale when sampling random action
                     rescaled_action = action
@@ -538,6 +544,7 @@ class SAC(OffPolicyRLModel):
             "policy": self.policy,
             "n_envs": self.n_envs,
             "action_noise": self.action_noise,
+            "random_exploration": self.random_exploration,
             "_vectorize_action": self._vectorize_action,
             "policy_kwargs": self.policy_kwargs
         }